+++ .github/workflows/breaking.yml
@@ -1,5 +1,5 @@
 # This script is provided by github.com/bool64/dev.
-name: gorelease
+name: breaking
 on:
   pull_request:
 env:
@@ -24,18 +24,18 @@ jobs:
             ~/go/bin/gorelease
           key: ${{ runner.os }}-gorelease
       - name: Install gorelease
-        run: |
-          test -e ~/go/bin/gorelease || go install golang.org/x/exp/cmd/gorelease@latest
+        run: test -e ~/go/bin/gorelease || go install golang.org/x/exp/cmd/gorelease@latest
       - name: Install egrep
-        run: |
-          which egrep || apt install egrep
+        run: which egrep || apt install egrep
       - name: Check API changes
         id: gorelease
         run: |
-          OUTPUT="$(gorelease || exit 0)"
-          OUTPUT="$(echo $OUTPUT | egrep -v '^.*from .*\/internal\/.* to .*\/internal\/.*$')"
-          echo "::set-output name=report::$OUTPUT"
-          exit 0
+          OUTPUT=$(gorelease || true);
+          OUTPUT=$(echo "$OUTPUT" | egrep -v "^.*from .*\/internal\/.* to .*\/internal\/.*$");
+          OUTPUT="${OUTPUT//'%'/'%25'}"
+          OUTPUT="${OUTPUT//$'\n'/'%0A'}"
+          OUTPUT="${OUTPUT//$'\r'/'%0D'}"
+          echo "::set-output name=report::$OUTPUT";
       - name: Comment Report
         if: always()
         uses: marocchino/sticky-pull-request-comment@v2
@@ -44,11 +44,6 @@ jobs:
           header: gorelease
           message: |
             ### API Changes
-
+            
             ${{ steps.gorelease.outputs.report }}
 
-      - name: Check breaking changes
-        run: |
-          if [[ ${{ steps.gorelease.outputs.report }} == *"Incompatible changes were detected"* ]]; then
-            exit 1
-          fi
